+++ deployment/ccip/view/solana/offramp.go
@@ -4,12 +4,14 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/ethereum/go-ethereum/common"
 	"github.com/gagliardetto/solana-go"
+	"github.com/mr-tron/base58"
+	chain_selectors "github.com/smartcontractkit/chain-selectors"
 	solOffRamp "github.com/smartcontractkit/chainlink-ccip/chains/solana/gobindings/ccip_offramp"
 	solState "github.com/smartcontractkit/chainlink-ccip/chains/solana/utils/state"
 
 	"github.com/smartcontractkit/chainlink/deployment"
-	"github.com/smartcontractkit/chainlink/deployment/ccip/view/shared"
 )
 
 type OffRampView struct {
@@ -74,12 +76,29 @@ func GenerateOffRampView(chain deployment.SolChain, program solana.PublicKey, re
 		if err = chain.GetAccountDataBorshInto(context.Background(), remoteChainPDA, &chainStateAccount); err != nil {
 			return view, fmt.Errorf("remote %d is not configured on solana chain %d", remote, chain.Selector)
 		}
+		onRamp := chainStateAccount.Config.OnRamp
 		view.SourceChains[remote] = OffRampSourceChainConfig{
 			IsEnabled:                 chainStateAccount.Config.IsEnabled,
 			IsRmnVerificationDisabled: chainStateAccount.Config.IsRmnVerificationDisabled,
 			LaneCodeVersion:           chainStateAccount.Config.LaneCodeVersion.String(),
-			OnRamp:                    shared.GetAddressFromBytes(remote, chainStateAccount.Config.OnRamp.Bytes[:]),
+			OnRamp:                    GetAddressFromBytes(remote, onRamp.Bytes[:onRamp.Len]),
 		}
 	}
 	return view, nil
+}
+
+func GetAddressFromBytes(chainSelector uint64, address []byte) string {
+	family, err := chain_selectors.GetSelectorFamily(chainSelector)
+	if err != nil {
+		return "invalid chain selector"
+	}
+
+	switch family {
+	case chain_selectors.FamilyEVM:
+		return common.BytesToAddress(address).Hex()
+	case chain_selectors.FamilySolana:
+		return base58.Encode(address)
+	default:
+		return "unsupported chain family"
+	}
 }
